[Feature/18] 챌린지 CRUD API 구현 (#21)
* dev 프로필 mariadb 사용 설정

* refactor: 챌린지와 챌린지 설명 테이블을 @ElementCollection을 이용해 일대다 관계로 매핑

* rename: content -> description

* refactor: 챌린지 CRUD API 정리

* .gitignore에 application-local.yml 추가

* refactor: 챌린지 전체 조회 URI 수정

+++ src/main/java/projectbuildup/mivv/domain/challenge/repository/ChallengeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import projectbuildup.mivv.domain.challenge.entity.Challenge;
 @Repository
-public interface ChallengeRepository extends JpaRepository<Challenge, Integer> {
+public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
 }
